Making the thing a helper instead of a ViewModel, finally
